+++ internal/ast/print.go
@@ -14,6 +14,7 @@ func Fprint(w io.Writer, prog *Program) error {
 	}
 	p := &printer{w: w}
 	p.line(`algoritmo "%s"`, prog.Name)
+	p.printConsole(prog.Console)
 	p.printDecls(prog.Consts, prog.Types, prog.Globals)
 	for _, sub := range prog.Subs {
 		p.line("")
@@ -48,6 +49,12 @@ func (p *printer) line(format string, args ...any) {
 		}
 	}
 	_, p.err = fmt.Fprintln(p.w)
+}
+
+func (p *printer) printConsole(settings []ConsoleStmt) {
+	for range settings {
+		p.line("dos")
+	}
 }
 
 func (p *printer) printDecls(consts []ConstDecl, types []TypeDecl, decls []VarDecl) {
@@ -96,6 +103,7 @@ func (p *printer) printSub(sub Subprogram) {
 	switch s := sub.(type) {
 	case *ProcedureDecl:
 		p.line("procedimento %s(%s)", s.Name.Text, paramsString(s.Params))
+		p.printConsole(s.Console)
 		p.printDecls(s.Consts, s.Types, s.Locals)
 		p.line("inicio")
 		p.indent++
@@ -104,6 +112,7 @@ func (p *printer) printSub(sub Subprogram) {
 		p.line("fimprocedimento")
 	case *FunctionDecl:
 		p.line("funcao %s(%s): %s", s.Name.Text, paramsString(s.Params), typeString(s.Return))
+		p.printConsole(s.Console)
 		p.printDecls(s.Consts, s.Types, s.Locals)
 		p.line("inicio")
 		p.indent++
@@ -121,6 +130,8 @@ func (p *printer) printStmts(stmts []Stmt) {
 
 func (p *printer) printStmt(stmt Stmt) {
 	switch s := stmt.(type) {
+	case *ConsoleStmt:
+		p.line("dos")
 	case *AssignStmt:
 		p.line("%s <- %s", exprString(s.Target), exprString(s.Value))
 	case *CallStmt: