+++ internal/ast/stmt.go
@@ -8,6 +8,12 @@ type Stmt interface {
 	Start() token.Pos
 }
 
+// ConsoleStmt requests console display in the program configuration section.
+type ConsoleStmt struct{ At token.Pos }
+
+func (*ConsoleStmt) stmtNode()          {}
+func (s *ConsoleStmt) Start() token.Pos { return s.At }
+
 // AssignStmt assigns to a variable or indexed vector element.
 type AssignStmt struct {
 	At     token.Pos